test(BA-7779): delete the first shape, now that nothing uses it
- `typed_scenario` and its runner are gone, and with them the deferred call, the
  actor-bound call, the accessor replayed to recover a field path, and every
  matcher that could look at one field.
- What survived moved to where it is used: reading what an adapter offers now
  lives beside the runner that reports it, and the timestamp condition beside the
  seeder that dates it.
- The rules say the three steps and the three verdicts a `then` may answer.
