+++ src/ai/backend/manager/api/adapters/session/report.md
@@ -12,21 +12,25 @@ Not exercised by any scenario: admin_search_kernels, batch_load_by_ids, batch_lo
 
 Given
 
-- 도메인 home-1
-- 도메인에 속한 사용자 한 명 준비
-  - 프로젝트 정책 default: 사용자를 만들 때 딸려 만들어지는 개인 프로젝트가 이 이름으로 찾는다
-  - 사용자 정책 user-policy-1: 사용자 한 명당 폴더 10개까지
-  - 키페어 정책 keypair-policy-1: 동시 세션 5개까지
-  - 슈퍼관리자 user-1: 자기 키와 개인 프로젝트를 갖는다
+- 도메인 하나와, 그 도메인에 속한 superadmin 한 명
+  - 도메인 host-1
+  - 도메인에 속한 사용자 한 명 준비
+    - 프로젝트 정책 default: 사용자를 만들 때 딸려 만들어지는 개인 프로젝트가 이 이름으로 찾는다
+    - 사용자 정책 user-policy-1: 사용자 한 명당 폴더 10개까지
+    - 키페어 정책 keypair-policy-1: 동시 세션 5개까지
+    - 슈퍼관리자 user-1: 자기 키와 개인 프로젝트를 갖는다
 
 When
 
-- SessionAdapter.admin_search — 슈퍼관리자 user-1
-  - AdminSearchSessionsInput()
+- SessionAdapter.admin_search — user-1이 필터 없이 전체 조회
 
 Then
 
-- 답 전체 일치
+- 답이 비어 있다
+  - items = []
+  - total_count = 0
+  - has_next_page = False
+  - has_previous_page = False
 
 #### [a-user-granted-nothing-may-not-search-sessions](/tests/scenario/bai_scenario/manager/session/test_session.py) — pass
 
@@ -34,19 +38,20 @@ Then
 
 Given
 
-- 도메인 home-1
-- 도메인에 속한 사용자 한 명 준비
-  - 프로젝트 정책 default: 사용자를 만들 때 딸려 만들어지는 개인 프로젝트가 이 이름으로 찾는다
-  - 사용자 정책 user-policy-1: 사용자 한 명당 폴더 10개까지
-  - 키페어 정책 keypair-policy-1: 동시 세션 5개까지
-  - 일반 사용자 user-1: 자기 키와 개인 프로젝트를 갖는다
+- 도메인 하나와, 그 도메인에 속한 user 한 명
+  - 도메인 host-1
+  - 도메인에 속한 사용자 한 명 준비
+    - 프로젝트 정책 default: 사용자를 만들 때 딸려 만들어지는 개인 프로젝트가 이 이름으로 찾는다
+    - 사용자 정책 user-policy-1: 사용자 한 명당 폴더 10개까지
+    - 키페어 정책 keypair-policy-1: 동시 세션 5개까지
+    - 일반 사용자 user-1: 자기 키와 개인 프로젝트를 갖는다
 
 When
 
-- SessionAdapter.admin_search — 일반 사용자 user-1
-  - AdminSearchSessionsInput()
+- SessionAdapter.admin_search — user-1이 필터 없이 전체 조회
 
 Then
 
-- 거부: NotEnoughPermission
+- 거부된다
+  - 거부: NotEnoughPermission
