+++ tests/scenario/AGENTS.md
@@ -15,24 +15,25 @@
 
 ## 한 행이 말하는 것
 
-시나리오는 다섯 부분을 갖는다. 각 부분은 뜻이 하나다.
+시나리오는 클래스 하나다. 이름과 보장하는 것, 그리고 세 단계를 답한다.
 
 | 부분 | 뜻 |
 |---|---|
 | `summary` | kebab 영문 요약. pytest 파라미터 id로도 쓰인다 |
-| `description` | 이 행이 무엇을 보장하는지 한 문장 |
-| `actor` | 누가 요청하는가. 시나리오가 심은 사용자 행 |
-| `given` | 요청 시점에 이미 참인 것 |
-| `when` / `then` | 요청과, 그 답 또는 예외 |
+| `describe` | 이 행이 무엇을 보장하는지 한 문장 |
+| `given` | 요청 시점에 이미 참이어야 하는 것을 세우고, 세운 것을 답한다 |
+| `when` | 그 값과 어댑터만 받아 호출한다 |
+| `then` | 심은 것과 답 또는 예외를 함께 받아 본다 |
 
-레포트는 이 다섯을 셋으로 묶어 찍는다. 심은 행이 `Given`, 요청이 `When`, 답 또는 거부가
-`Then`이다. 세 덩어리를 한 줄로 잇지 않는다. `When` 아래에는 그 호출이 실제로 받은 인자가
-하나씩 붙는다. 호출자가 손대지 않은 필드는 빠진다.
-
-- `description`은 생략할 수 없다. 상황과 검사하는 요구사항을 함께 적는다.
-- `description`에 식별자를 적지 않는다. 필드명, 예외 클래스, 파일 경로를 넣지 않는다.
-- `description`은 한글로 먼저 쓰고, 리뷰를 통과한 뒤 영문으로 옮긴다.
-- 문장이 표와 어긋나면 표가 아니라 문장을 먼저 의심한다. 문장이 의도이고 표가 구현이다.
+- `given`이 먼저 돌아 실제 값을 답하므로 값을 미뤄 두는 장치가 없다. 지연 호출도, 필드를
+  가리키는 람다도 쓰지 않는다.
+- `given`이 답하는 타입은 그 컴포넌트가 정한다. 뒤 단계가 그 타입을 그대로 받는다.
+- `when`은 `operation`으로 어느 어댑터 호출인지 말한다. 레포트가 그것으로 아무도 부르지 않은
+  호출을 센다.
+- 답이 노드를 감싸고 있으면 `when`이 벗겨서 준다. 그래야 여러 행이 검사기 하나를 같이 쓴다.
+- 설정을 바꿔 도는 행은 `Configured`를 함께 구현한다. 설정 경로를 적는 유일한 자리다.
+- `describe`는 한글로 먼저 쓰고, 리뷰를 통과한 뒤 영문으로 옮긴다.
+- `describe`에 식별자를 적지 않는다. 필드명, 예외 클래스, 파일 경로를 넣지 않는다.
 
 ## 행을 심는 법
 
@@ -88,11 +89,17 @@
 
 - 답의 일부만 보는 검사를 두지 않는다. 필드 하나만 보면 나머지가 조용히 바뀌어도 통과한다.
   그런 검사를 쓰기 쉽게 만드는 장치도 두지 않는다.
-- 답은 언제나 전체를 본다. 자리마다 셋 중 하나다 — 값으로 보거나, 조건으로 보거나, 이유를
-  적고 무시하거나. 어느 것도 아닌 자리는 없다.
-- 시나리오가 값을 말할 수 없는 필드만 빼낸다. 생성된 id는 이유를 적고 무시하며, 타임스탬프는
-  조건으로 본다.
-- 타임스탬프 조건의 기준은 `seed.since_started()`다. 임의의 시간 폭을 고르지 않는다.
+- 답은 언제나 전체를 본다. `then`이 자리마다 셋 중 하나를 답한다.
+
+  | 무엇 | 언제 | 레포트에 나오는 말 |
+  |---|---|---|
+  | `Same` | 시나리오가 값을 말할 수 있을 때 | `name = 'host-1'` |
+  | `Held` | 값은 말할 수 없고 조건만 걸 수 있을 때 | `created_at: 이 실행이 쓴 시각` |
+  | `Skipped` | 볼 수 없고 그 이유를 적을 수 있을 때 | `id: 무시함 — 데이터베이스가 만든다` |
+
+- 어느 것도 아닌 자리는 없다. 빠뜨린 자리는 조용히 통과하는 자리다.
+- `Held`가 거는 조건은 `Condition` 구현체다. 조건이 스스로 무엇인지 말하고, 그 말에 실제
+  값을 넣지 않는다. 넣으면 같은 동작인데도 실행마다 레포트가 달라진다.
 - 구현 디테일을 검사하지 않는다. 타입이 이미 보장하는 것을 다시 확인하지 않는다.
 
 ## 실패를 읽는 법
