+++ tests/scenario/bai_scenario/components/vfolder.py
@@ -7,21 +7,14 @@ from typing import Any, override
 
 from ai.backend.common.data.entity.user import UserID
 from ai.backend.common.data.entity.vfolder import VFolderEntityType
-from ai.backend.manager.api.adapters.vfolder.adapter import VFolderAdapter
-from ai.backend.manager.config.unified import ManagerUnifiedConfig
 from ai.backend.manager.data.domain.types import DomainData
 from ai.backend.manager.data.permission.types import Permission
 from ai.backend.manager.data.user.types import UserData
 from ai.backend.testutils.scenario_steps import Given
-from ai.backend.testutils.typed_scenario import (
-    TypedScenario,
-)
 from bai_scenario.components.domain import WAS_HERE, GrantedUser, SomeoneOf
 from bai_scenario.seeds.domain.domain import SeedDomain
 from bai_scenario.seeds.rbac.role import SeedPermission, SeedRole
 from bai_scenario.seeds.seeder import Laid, Seeder, SeedNest, SeedRow
-
-type VFolderScenario = TypedScenario[VFolderAdapter, ManagerUnifiedConfig]
 
 STORAGE_HOST = "local:volume1"
 """The one host the faked storage manager answers for."""
